+++ integration_tests/suite/test_dird_phonebook_backend.py
@@ -66,7 +66,6 @@ class TestPhonebookBackend(unittest.TestCase):
         self.phonebook_contact_crud = database.PhonebookContactCRUD(Session)
 
         self.phonebook = self.phonebook_crud.create(self.tenant_uuid, {'name': 'hogwarts'})
-        print(self.phonebook)
         contacts = [
             self.phonebook_contact_crud.create(self.tenant_uuid, self.phonebook['id'], c)
             for c in contact_bodies
@@ -83,7 +82,6 @@ class TestPhonebookBackend(unittest.TestCase):
         config = {
             'tenant_uuid': self.tenant_uuid,
             'name': 'hogwarts',
-            'db_uri': DB_URI,
             'phonebook_id': self.phonebook['id'],
             'searched_columns': ['firstname', 'lastname'],
             'first_matched_columns': ['number'],